block: Add driver to change the I/O scheduler when the screen turns off
This driver changes the I/O scheduler for a specified request queue to
Noop when the screen turns off in order to improve power consumption while
the screen is off, as well as reduce resume latency when the screen is
turned back on. After the screen is turned back on, the driver will change
the I/O scheduler back to what it originally was after a 10000ms delay.

Signed-off-by: Sultanxda <sultanxda@gmail.com> (reverted from commit 6004e9adb6b575123d6b119419d4c5b1d8533c79)